Add files via upload
ExcessEnthalpywithoutSMILES is the code used to create a neural network that did not use SMILES vectors to predict the excess enthalpy of the mixture. ExcessEnthalpywithSMILES includes a function that encodes SMILES vectors for a component and uses that as an input to the neural network. This was the method I recommended as future research of this work